% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\name{area_role}
\alias{area_role}
\title{Get the area role name for a poparray}
\usage{
area_role(x)
}
\arguments{
\item{x}{A poparray.}
}
\value{
A single character string naming the area dimension.
}
\description{
Get the area role name for a poparray
}
