% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{array_2_df}
\alias{array_2_df}
\title{Array to a Data Frame}
\usage{
array_2_df(arr, ..., stringAsFactors = FALSE, data_col = NULL)
}
\arguments{
\item{arr}{an array object}

\item{data_col}{a character string naming the column that will contain the data in the resulting data frame.
An explicit value overrides stored metadata. When omitted, the stored data-column
name is used when it is a single, non-missing, non-empty character string;
otherwise "value" is used.}
}
\value{
a data frame with the number of columns equal to the array dimensions + 1
}
\description{
Constructs a data frame from an array. The complimentary df_2_array() function does the opposite.  Note that
reconstruction of a data frame from an array is not perfect.  All dimension values in an array are stored as
character strings. For those dimension values that consist of numeric character strings (e.g., "123.45"), this
function will convert that dimension to a numeric column in the resulting data frame.  Likewise those that have
integer character values(e.g. 123) will be converted to integer.  All of the other values are changed to
character vectors or factors.
What this means is that a data .frame (df) with an an ordered factor will have an regular factor when arr <-
df_to_array(df) and then df_new <- arrary_to_df(arr). Therefore df may not equal df_new
}
\examples{
# Convert an array to a data frame
arr <- array(
  c(100, 95, 102, 98),
  dim = c(2, 2),
  dimnames = list(year = c("2020", "2021"), sex = c("Female", "Male"))
)
attr(arr, "data_col") <- "population"

df <- array_2_df(arr)
head(df)

}
\keyword{internal}
