% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_class.r
\name{as.data.frame.pop_projection}
\alias{as.data.frame.pop_projection}
\alias{as_tibble.pop_projection}
\title{Corerce pop_projection to a data frame}
\usage{
\method{as.data.frame}{pop_projection}(x, ..., include_level = TRUE, include_model = TRUE)

\method{as_tibble}{pop_projection}(x, ..., include_level = TRUE, include_model = TRUE)
}
\arguments{
\item{x}{a pop_projection object}

\item{...}{}

\item{include_level}{the default TRUE means add a column with the confidence level used for the projection}

\item{include_model}{the default TRUE causes a column has the model used fo the projection}
}
\value{
a data frame for as.data.frame() and tibble for as_tibble()
}
\description{
Transforms the  poparray in x to a data frame or tibble.  Causes a realization of all data in the lower, upper,
and  population_project cubes.
}
\examples{
# TO DO
}
