% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_class.r
\name{as.poparray.pop_projection}
\alias{as.poparray.pop_projection}
\title{Coerce a poparray_projection to a poparray}
\usage{
\method{as.poparray}{pop_projection}(x, which = c("projection", "std_error"), ...)
}
\arguments{
\item{x}{a pop_projection object}

\item{which}{whn x is pop_projection object, you choose which cube to make into a single poparray:
*  projected
*  lower
*  upper}

\item{...}{}
}
\description{
at pop_projection consists of three poparray cubes, this coercion returns a single poparray object
based on the which argument.
}
