% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\name{as.poparray}
\alias{as.poparray}
\title{Coerce to a poparray Object}
\usage{
as.poparray(x, data_col = "population", ...)
}
\arguments{
\item{x}{object to coerce like a data frame or a table/array.}

\item{data_col}{single character string with the population column name.}

\item{...}{additional arguments to pass (see methods).}
}
\value{
a poparray classed object
}
\description{
Used to coerce a data frame or an array/table to a poparray object.
}
