% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr_pop_class.r
\name{as.tarr_pop}
\alias{as.tarr_pop}
\title{Coerce to a tarr_pop Object}
\usage{
as.tarr_pop(x, data_col = "population", ...)
}
\arguments{
\item{x}{object to coerce like a data frame or a table/array.}

\item{data_col}{single character string with the population column name.}

\item{...}{additional arguments to pass (see methods).}
}
\value{
a tarr_pop classed object
}
\description{
Used to coerce a data frame or an array/table to a tarr_pop object.
}
