% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_class.r
\name{as.tarr_pop.tarr_projection}
\alias{as.tarr_pop.tarr_projection}
\title{Coerce a tarr_projection to a tarr_pop}
\usage{
as.tarr_pop.tarr_projection(x, which = c("projected", "lower", "upper"), ...)
}
\arguments{
\item{x}{a tarr_projection object}

\item{which}{whn x is tarr_projection object, you choose which cube to make intoa single tarr_pop:
*  projected
*  lower
*  upper}

\item{...}{}
}
\description{
at tarr_projection consists of three tarr_pop cubes, this coercion returns a single tarr_pop object
based on the which argument.
}
