% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr_pop_class.r
\name{as_tibble.tarr_pop}
\alias{as_tibble.tarr_pop}
\title{Coerce to a Tibble}
\usage{
\method{as_tibble}{tarr_pop}(
  x,
  stringsAsFactors = TRUE,
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{x}{tarr_pop object}

\item{stringsAsFactors}{Logical, passed to as.data.frame.tarr_pop()}

\item{...}{Not used}

\item{.name_repair}{Name repair strategy for tibble}
}
\value{
a tibble
}
\description{
S3 method for class 'tarr_pop'.
Coercion is done via as.data.frame.tarr_pop for consistency.
}
