% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{assign_array_levels}
\alias{assign_array_levels}
\title{Assign new values to a dimension}
\usage{
assign_array_levels(arr, dimname, char_vec)
}
\arguments{
\item{arr}{is the array to receive the new value}

\item{dimname}{the dimension name}

\item{char_vec}{is the character vector to assign to dimension}
}
\value{
the modified array
}
\description{
Assign a new character vector a levels to a dimension.  Missing values from the current levels are dropped before
assigning the new values.
}
