% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr_pop_class.r
\name{autoplot.tarr_pop}
\alias{autoplot.tarr_pop}
\title{Autoplot a tarr_pop}
\usage{
\method{autoplot}{tarr_pop}(
  object,
  ...,
  proportional = FALSE,
  fill = c("royalblue2", "palevioletred3"),
  source = attr(object, "source")[["note"]]
)
}
\arguments{
\item{object}{a tarr_pop object}

\item{...}{additional arguments}

\item{proportional}{logical; if TRUE plot proportions instead of counts}

\item{fill}{named vector of colors for male/female}

\item{source}{caption for plot, defaults to tarr_pop source attribute}
}
\value{
a ggplot
}
\description{
Method for autoplot() to chart a tarr_pop.  If gender is available, will
create a population pyramid, otherwise a line chart by age.
}
