% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\name{data_col}
\alias{data_col}
\alias{data_col<-}
\title{Get or set the name of the poparray data column}
\usage{
data_col(x)

data_col(x, values) <- value
}
\arguments{
\item{x}{A poparray object.}
}
\value{
Character string.
}
\description{
poparray objects have a \code{data_col} attribute that is used to name the column that holds the numeric data when
coercing to a data frame \code{\link[=as.data.frame]{as.data.frame()}} or tibble \code{\link[tibble:as_tibble]{tibble::as_tibble()}}. This convenience function can retrieve or
set that attribute.
}
