% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{df_2_array}
\alias{df_2_array}
\title{Data Frame to Array}
\usage{
df_2_array(df, data_col = "value")
}
\arguments{
\item{df}{a data frame with one column that will act as the data in an array.}

\item{data_col}{a single character string containing the column name for the source of data used to construct the
array.}
}
\value{
an array with dimensions ncol() - 1. The array will have the "data_col" attribute, a character string with the
original name of the column that contained the data.
}
\description{
A data frame can be made into an array with one column serving as the source of data, and the other columns becoming
dimensions. This function takes a data frame and returns an array with dimnames and values. The input must already
contain exactly one row for every Cartesian combination of the dimension labels; incomplete or duplicate
dimensional coordinates are errors.
Generally, arrays are a more efficient means of storing data compared to data frames. Using the indexing operator to
return subsets is very fast. This function is used internally, but is exposed for use outside the package.  The
complimentary array_2_df() is the reverse of this function.
}
\examples{
# convert a data frame into an array
df <- data.frame(
  year = c("2020", "2020", "2021", "2021"),
  sex = c("Female", "Male", "Female", "Male"),
  population = c(100, 95, 102, 98)
)

arr <- df_2_array(df, data_col = "population")

dim(arr)
dimnames(arr)

}
\keyword{internal}
