% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{drop_array_level}
\alias{drop_array_level}
\title{Drop Level(s) from an Array Dimension
Internal function to handle dropping one or more levels from an array dimension}
\usage{
drop_array_level(arr, dim_name, drop_values)
}
\arguments{
\item{arr}{an array}

\item{dim_name}{is the name of the dimension to drop the level from}

\item{drop_values}{are the specific values to drop}
}
\value{
the modified array
}
\description{
Drop Level(s) from an Array Dimension
Internal function to handle dropping one or more levels from an array dimension
}
\examples{
tmp <- drop_array_level(census, "age.char", "All")
}
