% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_source}
\alias{get_source}
\title{Retrieve the poparray source}
\usage{
get_source(obj)
}
\arguments{
\item{x}{is the object with a source attribute, like a poparray.}
}
\value{
A named character list with entries such as \code{note}, \code{source}, and \code{updated}.
}
\description{
Gets the list of information contained in the \code{source} attribute of \code{x}. Objects from poparray have source attributes
as possibly other objects.
}
\keyword{internal}
