% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pop_array.r
\name{h5_has_cube_schema}
\alias{h5_has_cube_schema}
\title{Check if a file has the cube metadata schema}
\usage{
h5_has_cube_schema(path, info = NULL)
}
\arguments{
\item{path}{HDF5 file path.}
}
\value{
Logical scalar.
}
\description{
Check if a file has the cube metadata schema
}
\keyword{internal}
