% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pop_array.r
\name{h5_read_scalar_chr}
\alias{h5_read_scalar_chr}
\title{Read a scalar metadata value from HDF5}
\usage{
h5_read_scalar_chr(path, name)
}
\arguments{
\item{path}{HDF5 file path.}

\item{name}{HDF5 dataset path.}
}
\value{
Length-1 character scalar.
}
\description{
Read a scalar metadata value from HDF5
}
\keyword{internal}
