% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pop_array.r
\name{h5_read_scalar_chr_if_present}
\alias{h5_read_scalar_chr_if_present}
\title{Read scalar metadata if present}
\usage{
h5_read_scalar_chr_if_present(path, name)
}
\arguments{
\item{path}{HDF5 file path.}

\item{name}{HDF5 dataset path.}
}
\value{
Length-1 character scalar or NULL.
}
\description{
Read scalar metadata if present
}
\keyword{internal}
