% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{label_factory}
\alias{label_factory}
\title{Factory function to generate wrapper functions for retrieving specific dimension labels from poparray objects}
\usage{
label_factory(dimension)
}
\arguments{
\item{dimension}{should be a valid dimension of the array}
}
\description{
Factory function to generate wrapper functions for retrieving specific dimension labels from poparray objects
}
