% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr_pop_class.r
\name{make_pyramid}
\alias{make_pyramid}
\title{Make a population pyramid}
\usage{
make_pyramid(
  df,
  xlim = NULL,
  xaxis = FALSE,
  yaxis = FALSE,
  area_label = FALSE,
  year_label = FALSE,
  col = c(Male = "royalblue3", Female = "palevioletred3")
)
}
\arguments{
\item{df}{a data frame converted from a tarr_pop object. Expects one year and
one area.}

\item{xlim}{numeric scalar with the max abs limit for x-axis.}

\item{xaxis}{logical; draw x-axis if TRUE}

\item{yaxis}{logical; draw y-axis labels if TRUE}

\item{area_label}{logical; show area label}

\item{year_label}{logical; show year label}

\item{col}{named vector of colors for male and female bars.}
}
\value{
plots the pyramid
}
\description{
Takes a data frame with population figures with area, year, sex and ages
then creates a population pyramid. Internally used, not exported.
}
