% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr_pop_class.r
\name{new_tarr_pop}
\alias{new_tarr_pop}
\alias{is.tarr_pop}
\title{Create tarr_pop objects}
\usage{
new_tarr_pop(
  x,
  dimnames_list = dimnames(x),
  data_col = "population",
  source = NULL,
  ...
)

is.tarr_pop(x)
}
\arguments{
\item{x}{object to check type or convert}

\item{dimnames_list}{Named list of dimnames, with names: \code{year}, \code{area.name}, \code{sex}, \code{age.char}, \code{race}, \code{ethnicity}.}

\item{data_col}{Name of the data column when converted to a data.frame/tibble.}

\item{source}{Optional metadata describing data source (list or character).}

\item{...}{Reserved for future use.}
}
\value{
An object of class 'tarr_pop'.
}
\description{
Create or test for objects of type "tarr_pop".

Checks that the object is of class tarr_pop
}
\details{
Constructor for tarr_pop objects. The S3 class 'tarr_pop' wraps a DelayedArray / HDF5Array The underlying data is
held in \code{handle}, and standardized dimnames are kept in  \code{dimn}. Attributes are used for 'age_iv', 'data_col', and
'source'.
}
