% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_semantics.r
\name{pa_is_interval}
\alias{pa_is_interval}
\title{Predicate: interval scale}
\usage{
pa_is_interval(sem)
}
\arguments{
\item{sem}{Object to check.}
}
\value{
\code{TRUE} when \code{sem} is \code{DimSemantics} with \code{scale_type == "interval"}.
}
\description{
Predicate: interval scale
}
\keyword{internal}
