% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_semantics.r
\name{pa_is_partition}
\alias{pa_is_partition}
\title{Predicate: partition partition type}
\usage{
pa_is_partition(sem)
}
\arguments{
\item{sem}{Object to check.}
}
\value{
\code{TRUE} when \code{sem} is \code{DimSemantics} with
\code{partition_type == "partition"}.
}
\description{
Predicate: partition partition type
}
\keyword{internal}
