% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_semantics.r
\name{pa_is_set}
\alias{pa_is_set}
\title{Predicate: set partition type}
\usage{
pa_is_set(sem)
}
\arguments{
\item{sem}{Object to check.}
}
\value{
\code{TRUE} when \code{sem} is \code{DimSemantics} with \code{partition_type == "set"}.
}
\description{
Predicate: set partition type
}
\keyword{internal}
