% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_semantics.r
\name{pa_update_dim_semantics}
\alias{pa_update_dim_semantics}
\title{Update a DimSemantics object}
\usage{
pa_update_dim_semantics(sem, ...)
}
\arguments{
\item{sem}{A \code{DimSemantics} object.}

\item{...}{Named properties to update.}
}
\value{
Updated \code{DimSemantics} object.
}
\description{
Internal controlled updater for \code{DimSemantics} properties.
}
\keyword{internal}
