% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr_pop_class.r
\name{plot.tarr_pop}
\alias{plot.tarr_pop}
\title{Plot Method for tarr_pop Object}
\usage{
\method{plot}{tarr_pop}(x, ...)
}
\arguments{
\item{x}{a tarr_pop object}

\item{...}{not used}
}
\value{
plots the data without returning anything
}
\description{
Plots a population pyramid based on a tarr_pop
}
