% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_class.r
\docType{class}
\name{pop_projection}
\alias{pop_projection}
\title{Projection class}
\description{
An S3 class representing a time-based projection. The class is a wrapper around an HDF5Array that contains the
projected population and a dimension of the error. Objects of this class are typically created by \code{\link[=project]{project()}}.
}
\section{Structure}{

A \code{pop_projection} object is a list with a HDF5Array cube as a member. The cube is very similar to a poparray but
wheres athe poparray figures are for population, these are projected population figures and another set of figures
are added that represents the standard error and labeled as "stderr".

Attributes included:
\itemize{
\item \strong{level} is the confidence level used
\item \strong{methods_used} are the unique time series forecasting methods used across all cells
\item \strong{n_fallback} are the number of cells that required fallback as the forecasting method .
\item \strong{source} is the  "projected from" plus the original poparray source
\item \strong{base_years} are the base years or time unit used to project/forecast
\item \strong{created} is the date and time stamp at time the object was created.
}
}

\seealso{
\itemize{
\item \code{\link[=project]{project()}} to create a projection.
\item \code{\link[=as.poparray.pop_projection]{as.poparray.pop_projection()}} to select one of the poparray objects
}
}
\keyword{internal}
