% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\docType{class}
\name{poparray}
\alias{poparray}
\title{Poparray class}
\description{
S4 class for population arrays or often referred to as cubes. \code{poparray} extends the Bioconductor \code{DelayedArray} and
adds explicit role and metadata slots. Memory is saved and speed increased by opening a poparray cube from disk and
manipulating the dimensions/metadata before realizing (loading) the selected array/cube in memory.  Typical work flow is
to open the cube via the \code{\link[=open_poparray]{open_poparray()}} function, filtering the dimension levels that are needed through the \code{[}
index operator or filter() function.  Dimension levels can also be collapsed using \code{\link[=collapse_dim]{collapse_dim()}}.  The cube can
then be realized \code{\link[=as.array]{as.array()}}, \code{\link[=as.data.frame]{as.data.frame()}}, or \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.
}
\section{Slots}{

\describe{
\item{\code{time_role}}{Name of the time dimension.}

\item{\code{area_role}}{Name of the area dimension.}

\item{\code{strata_roles}}{Character vector of optional stratification dimensions.}

\item{\code{dim_semantics}}{Named list of \code{DimSemantics} objects (one per dimension).}

\item{\code{data_col}}{Name used for value column in tabular coercions.}

\item{\code{source}}{Provenance metadata as a named list.}
}}

