% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pop_array.r
\name{resolve_extdata_dir}
\alias{resolve_extdata_dir}
\title{Resolve extdata directory for this package}
\usage{
resolve_extdata_dir(strict = TRUE)
}
\arguments{
\item{strict}{Logical; error if the directory cannot be located.}
}
\value{
Absolute path to extdata directory.
}
\description{
Resolve extdata directory for this package
}
\keyword{internal}
