% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{set_source_url}
\alias{set_source_url}
\title{Set source attributes}
\usage{
set_source_url(obj, nm, url, pop_type = NULL)
}
\arguments{
\item{obj}{the object to set the source attributes for}

\item{nm}{the name of the source, e.g. "Texas Demographic Center"}

\item{url}{the url of the source, e.g. "https://demographics.texas.gov"}
}
\value{
the object with the source attribute set to include name of source,
url, and the date the object was last created/revised set to today()
}
\description{
Attributes like the source program, url, and date created.
}
