% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr_pop_class.r
\name{split.tarr_pop}
\alias{split.tarr_pop}
\title{Split Population}
\usage{
\method{split}{tarr_pop}(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{a tarr_pop object}

\item{f}{the dimension name to use as the splitting variable. Must be a
single character string.}

\item{drop}{if the tarr_pop array should drop the dimension being split.
The default is FALSE.}

\item{...}{arguments to pass to other functions, not used.}
}
\value{
a list of tarr_pop
}
\description{
Splits a tarr_pop object by dimension values.
}
