% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_class.r
\name{std_error}
\alias{std_error}
\title{Extract standard errors from a poparray_projection}
\usage{
std_error(x)
}
\arguments{
\item{x}{A \code{poparray_projection} object.}
}
\value{
A \code{DelayedArray} slice retaining all dimensions, with \code{stat}
restricted to \code{"std_error"} (\code{drop = FALSE}).
}
\description{
Returns the delayed array slice where the \code{stat} dimension is
\code{"std_error"}.
}
