% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr.pop-package.R
\docType{package}
\name{tarr.pop}
\alias{tarr.pop-package}
\alias{tarr.pop}
\title{tarr.pop-package.R}
\description{
The package provides a flexible framework for storing, managing, and analyzing population data using multidimensional
arrays. It is designed to support demographic and epidemiologic workflows where population counts are organized
across dimensions like time, geography, age, sex, race, ethnicity, or other descriptors like income groups, home
values, household size, et cetera.

The package enables users to work with population data from multiple sources in a consistent structure, preserving
important semantic information about each dimension (e.g., whether categories are mutually exclusive or overlapping).
This allows for safe aggregation, filtering, and transformation of population data while maintaining reproducibility
and interpretability.

Population data are stored using  \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} and \link[HDF5Array:HDF5Array-class]{HDF5Array::HDF5Array} disk-backed arrays, allowing users to work with
large data sets that exceed available memory. Operations such as filtering, sub-setting, and projection are performed
lazily, minimizing memory usage and improving performance by avoiding unnecessary data realization.

The package also provides tools for extending population cubes, including time-series projection methods and a
standardized HDF5-based storage format with embedded metadata. Refer to the \link{population} help topic for available
example data sets and their formats.
}
\references{
Population figure sources are:
\itemize{
\item Decennial censuses and Estimates \href{https://www.census.gov/}{U.S. Census Bureau}
\item \href{https://demographics.texas.gov/Data/TPEPP/Estimates/}{Texas Demographic Center Estimates program}
\item \href{https://demographics.texas.gov/Data/TPEPP/Projections/}{Texas Demographic Center Projections program}
\item \href{https://seer.cancer.gov/data-software/uspopulations.html}{National Cancer Institute SEER Program}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RussJones007/tarr.pop}
  \item Report bugs at \url{https://github.com/RussJones007/tarr.pop/issues}
}

}
\author{
Russ Jones \href{mailto:RussJones007@gmail.com}{RussJones007@gmail.com}
}
\keyword{package}
