% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_class.r
\docType{class}
\name{tarr_projection}
\alias{tarr_projection}
\title{Projection object}
\description{
An S3 class representing a time-based projection.

Objects of this class are typically created by \code{\link[=project]{project()}}.  It contains the predicted population for each cell
in the original matrix .
}
\section{Structure}{

A \code{tarr_projection} object is a list with three tarr_pop cubes:
\itemize{
\item lower: is the lower limits of the calculated confidence interval.
\item projected: The projected population figures.
\item upper: is the upper limit of the calculated confidenceinterval.
}

Attributes included:
\itemize{
\item \strong{level} is the confidence level used
\item \strong{methods_used} are the unique methods actually used across all cells
\item \strong{n_fallback} are the number of cells that required fallback as the forecasting method .
\item \strong{source} is the  "projected from" plus the original tarr_pop source
\item \strong{base_years} are the base years used to project/forecast
\item \strong{created} is the date and time stamp at time the object was created.
}
}

\seealso{
\itemize{
\item \code{\link[=project]{project()}} to create a projection.
\item \code{\link[=as.tarr_pop.tarr_projection]{as.tarr_pop.tarr_projection()}} to select one of the tarr_pop objects
}
}
\keyword{internal}
