% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\name{time_role}
\alias{time_role}
\title{Get the time role name for a poparray}
\usage{
time_role(x)
}
\arguments{
\item{x}{A poparray.}
}
\value{
A single character string naming the time dimension.
}
\description{
Get the time role name for a poparray
}
