% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\name{validate_poparray}
\alias{validate_poparray}
\title{Validate poparray object}
\usage{
validate_poparray(x)
}
\arguments{
\item{x}{a poparray object to validte}
}
\value{
x, otherwise throws an error
}
\description{
Validate poparray object
}
\keyword{internal}
