% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\name{warn_if_realization_large}
\alias{warn_if_realization_large}
\title{Warn for large realized array}
\usage{
warn_if_realization_large(x, bytes_threshold = 5e+07 * 8)
}
\arguments{
\item{x}{A poparray.}

\item{bytes_threshold}{Threshold in bytes for warning (default ~400 MB).}
}
\value{
x invisibly.
}
\description{
Used for the side effect of issuing a warning when a realized array is very large.
}
\keyword{internal}
