% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\name{warn_if_realization_large_delayed}
\alias{warn_if_realization_large_delayed}
\title{Warn for large delayed realizations}
\usage{
warn_if_realization_large_delayed(
  x,
  bytes_threshold = 5e+07 * 8,
  what = "Coercing this array"
)
}
\arguments{
\item{x}{A \code{DelayedArray}.}

\item{bytes_threshold}{Threshold in bytes for warning (default ~400 MB).}

\item{what}{Short label describing the eager action being taken.}
}
\value{
x invisibly.
}
\description{
Used for the side effect of issuing a warning when realizing a delayed array
would be large.
}
\keyword{internal}
