% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_class.r
\name{as.data.frame.poparray_projection}
\alias{as.data.frame.poparray_projection}
\alias{as_tibble.poparray_projection}
\title{Coerce poparray_projection to a data frame}
\usage{
\method{as.data.frame}{poparray_projection}(x, ..., include_level = TRUE, include_model = TRUE)

\method{as_tibble}{poparray_projection}(x, ..., include_level = TRUE, include_model = TRUE)
}
\arguments{
\item{x}{a poparray_projection object}

\item{...}{}

\item{include_level}{the default TRUE means add a column with the confidence level used for the projection}

\item{include_model}{the default TRUE causes a column has the model used fo the projection}
}
\value{
a data frame for as.data.frame() and tibble for as_tibble()
}
\description{
Transforms the delayed array in \code{x} to a data frame or tibble. This is an eager
realization. The output contains \code{projection} and \code{std_error} columns.
}
\examples{
# TO DO
}
