% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dim_labels}
\alias{dim_labels}
\alias{assign_label_factory}
\alias{ages}
\alias{races}
\alias{ethnicities}
\alias{sexes}
\alias{areas}
\alias{years}
\title{Dimension Labels}
\usage{
dim_labels(arr, dimname, remove = NULL)

assign_label_factory(dim_name)

ages(arr, ...)

races(arr, ...)

ethnicities(arr, ...)

sexes(arr, ...)

areas(arr, ...)

years(arr, ...)
}
\arguments{
\item{arr}{A poparray object.}

\item{dimname}{the name of the dimension to get the labels return}

\item{remove}{is a character vector which will be removed when returning the labels for a dimension.
Alternatively this can be a regex expression created by the \code{\link[stringr:regex]{stringr::regex()}} function}

\item{...}{used by the wrapper functions to pass arguments to the dim_labels() function.}

\item{dimension}{is the name of the dimension}
}
\value{
a character vector of labels for the chosen dimension#'

a function to assign x (an array), levels passed as values, to the specified dimension name
}
\description{
Retrieves the labels for an array dimension with the option to remove one or more labels.
These functions are useful when sub-setting arrays, especially when choosing elements from a poparray.
}
\details{
Simple wrapper functions around dim_label() to retrieve and assign each dimension label values that are included in a
poparray object.
\itemize{
\item ages() the age.char dimension, ages(arr) <- values
\item sexes() the sex dimension
\item races() the race dimension
\item ethnicities() the ethnicity dimension
\item areas() the areas (counties) dimension
\item years() the years dimension. Note that this conflicts with the same function name from the lubridate package.
The best practice is to use tarr.pop::years()
}
}
\examples{

selected_ages <- dim_labels(census, "age.char", remove = c("All", "85 +"))
selected_ages
cen <- census[year = "2020",
             area.name = "Tarrant",
             sex = "All",
             age.char = selected_ages,
             race = "All",
             ethnicity = "All",
             drop = FALSE]

dimnames(cen)["age.char"]

dim_labels(arr = cen, dimname = "age.char", remove =  "All")

# or using ages() without removing any age group
ages(cen)
dim_labels(cen, dimname = "race", remove = stringr::regex(combination))

# to do
}
