% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\name{dimnames,poparray-method}
\alias{dimnames,poparray-method}
\title{dimnames method for poparry}
\usage{
\S4method{dimnames}{poparray}(x)
}
\value{
the poparray
}
\description{
The dimnames() function is time expensive.  This method caches the dimension names
to prevent pinging the on disk file.
}
