% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_source}
\alias{get_source}
\title{Retrieve the tarr_pop source}
\usage{
get_source(obj)
}
\arguments{
\item{x}{is the object with a source attribute, like a tarr_pop array.}
}
\value{
a named list of character  - names are note, source and updated.
}
\description{
Gets the list of information contained in "source" attribute of 'x'.  Objects from tarr_pop have a source attributes
as possibly other objects.
}
\keyword{internal}
