% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ages.R
\name{group_ages}
\alias{group_ages}
\title{Ages into Groups}
\usage{
group_ages(pop, age_groups, dimension_name = "age.char")
}
\arguments{
\item{pop}{A poparray object with an age dimension name or column.}

\item{age_groups}{The desired age groups to use, this can be a vector that \code{\link[rage:as.age_group]{rage::as.age_group()}} accepts such
as a numeric or character vector.  Character vectors must conform to the format expected  by \code{\link[rage:as.age_group]{rage::as.age_group()}}}

\item{dimension_name}{The dimension or column name that contains the age data.  "age.char" is the default.}
}
\value{
The object type as in the pop argument, age groups collapsed and the count data (population) summed.
}
\description{
The ages dimension or column name is set to age groups and the count data (population) is summed in the new
returned object.
}
\examples{
# Load a poparray, subset lazily, then group ages.
# census <- open_poparray(population$census.bureau$census)
# census <- dplyr::filter(census, year \%in\% c("2010", "2020"))
# census_age_grouped <- group_ages(census, rage::age_groups$ILI)
}
