% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ages.R
\name{group_ages}
\alias{group_ages}
\title{Ages into Groups}
\usage{
group_ages(pop, age_groups, dimension_name = "age.char")
}
\arguments{
\item{pop}{a tarr_pop (array) object with an age dimension name or column}

\item{age_groups}{The desired age groups to use, this can be a vector that \code{\link[rage:as.age_group]{rage::as.age_group()}} accepts such
as a numeric or character vector.  Character vectors must conform to the format expected  by \code{\link[rage:as.age_group]{rage::as.age_group()}}}

\item{dimension_name}{The dimension or column name that contains the age data.  "age.char" is the default.}
}
\value{
The object type as in the pop argument, age groups collapsed and the count data (population) summed.
}
\description{
The ages dimension or column name is set to age groups and the count data (population) is summed in the new
returned object.
}
\examples{
census <- county_population(
     population$census.bureau$census(),
     year = c("2010", "2020"),
     county = "Tarrant",
     sex = c("Female", "Male"),
     age  =  c(as.character(0:99), "100-104"),
     race = "All",
     ethnicity = c("Hispanic", "Non-Hispanic"),
     drop = FALSE)

census_age_grouped <- group_ages(pop = census, rage::age.groups$ILI)
}
