% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ages.R
\name{group_array_by_levels}
\alias{group_array_by_levels}
\title{Group an Array Dimension}
\usage{
group_array_by_levels(
  arr,
  dim_name,
  groups,
  keep_unmapped = c("error", "drop", "keep"),
  other_label = "Other"
)
}
\arguments{
\item{arr}{A poparray object or similarly structured array-like object.}

\item{dim_name}{Name of dimension to group}

\item{groups}{Group specification which maps the new levels to those levels being collapsed. This can be a named list
with the name of the group and values being what will be in the group, or a named vector mapping each new group to
an individual level.}

\item{keep_unmapped}{a character vector of length one flagging on what happens when not all the levels are included
in groups.  The values this argument can be:
"error"  - throw an error and stop if all the original levels are not included.
"drop"   - exclude un-mapped levels entirely.
"keep"   - keep un-mapped levels with their original labels.
"other"  - all un-mapped levels are summarized and placed under the other_label.
@param other_label is where any levels not included in groups, but "other" is set, are summed into this category.
"Other" is the default level name.}
}
\value{
Object with grouped dimension and summed counts
}
\description{
Groups specified dimension into categories and sums population counts.
}
\examples{
# Group age dimension
# Coming soon
}
