% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tarr_pop.r
\name{open_poparray}
\alias{open_poparray}
\title{Open population series}
\usage{
open_poparray(series_id, dataset = "/pop", data_col = "population")
}
\arguments{
\item{series_id}{Name of the population series.}

\item{dataset}{Path to the data in the HDF5 file.}

\item{data_col}{Name of the value column when coercing to a data frame.}
}
\value{
A poparray
}
\description{
Using the name of a population series that exists on disk, open it and return
as a poparray object.
}
