% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tarr_pop.r
\name{open_tarr_pop}
\alias{open_tarr_pop}
\title{Open population series}
\usage{
open_tarr_pop(series_id, dataset = "/pop", data_col = "population")
}
\arguments{
\item{series_id}{name of the population series.  See the population list variable for easier selection in the IDE.}

\item{dataset}{path to the data in the HDF5Array file}

\item{data_col}{name of the column with the population figures when array is transoformed into a data frame.}
}
\value{
the selected population series as a tarr_pop object
}
\description{
Using the name of a population series that exists on disk,  open it and retunr as a tarr_pop object.
}
\examples{
#TBD
}
