% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poparray_class.r
\name{polish_df}
\alias{polish_df}
\title{Polish data frames after coercing}
\usage{
polish_df(df, stringsAsFactors = TRUE, time_dim = "year")
}
\arguments{
\item{df}{}

\item{stringsAsFactors}{}

\item{time_dim}{}
}
\value{
a data frame
}
\description{
Used by as.data.frame() as a suport function
}
\examples{
#todo
}
\keyword{internal}
