% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr_pop_class.r
\name{polish_df}
\alias{polish_df}
\title{Finish polishing df}
\usage{
polish_df(df, stringsAsFactors = TRUE)
}
\arguments{
\item{df}{a data frame object, or tibble or data.table.}

\item{stringsAsFactors}{Logical, currently not used for factor conversion, but
kept for compatibility.}
}
\value{
the same object as df but with appropriate types / ordered factors.
}
\description{
"Polishes" the df argument from as.data.frame.tarr_pop or as_tibble.tarr_pop.
In future, may also be called by as.data.table.tarr_pop.
}
\keyword{internal}
