% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pop_array.r
\name{read_dimnames_from_cube}
\alias{read_dimnames_from_cube}
\title{Read dimnames metadata from migrated cube}
\usage{
read_dimnames_from_cube(path)
}
\arguments{
\item{path}{HDF5 file path.}
}
\value{
Named list of character vectors.
}
\description{
Read dimnames metadata from migrated cube
}
\keyword{internal}
