% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pop_array.r
\name{read_registry_row}
\alias{read_registry_row}
\title{Read registry row from one migrated cube file}
\usage{
read_registry_row(path)
}
\arguments{
\item{path}{HDF5 file path.}
}
\value{
One-row data.frame.
}
\description{
Read registry row from one migrated cube file
}
\keyword{internal}
