% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pop_array.r
\name{read_roles_from_cube}
\alias{read_roles_from_cube}
\title{Read dimension roles metadata from migrated cube}
\usage{
read_roles_from_cube(path)
}
\arguments{
\item{path}{HDF5 file path.}
}
\value{
Named list with time, area, strata.
}
\description{
Read dimension roles metadata from migrated cube
}
\keyword{internal}
