% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pop_array.r
\name{read_source_from_cube}
\alias{read_source_from_cube}
\title{Read source metadata from migrated cube}
\usage{
read_source_from_cube(path)
}
\arguments{
\item{path}{HDF5 file path.}
}
\value{
Named character vector with source fields.
}
\description{
Read source metadata from migrated cube
}
\keyword{internal}
