% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarr.pop-package.R
\docType{package}
\name{tarr.pop}
\alias{tarr.pop-package}
\alias{tarr.pop}
\title{tarr.pop-package.R}
\description{
Contains the population figures for Texas, and all counties. The population data sets contain age, sex, race, and
ethnicity (asre) data. The estimated population by Zip Code Tabulation Areas (zcta) are also available.  Refer to the
population help topic for what is contained and data format for the data sets.
}
\section{Package functions}{

\itemize{
\item county_population() for convenient access to the county asre data sets for years of interest.
\item retrieve_zip_code_population() to select different years of zip codes and associated population estimates for Tarrant
County.
}
}

\references{
Population figure sources are:
\itemize{
\item Decennial censuses and Estimates \href{https://www.census.gov/}{U.S. Census Bureau}
\item \href{https://demographics.texas.gov/Data/TPEPP/Estimates/}{Texas Demographic Center Estimates program}
\item \href{https://demographics.texas.gov/Data/TPEPP/Projections/}{Texas Demographic Center Projections program}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RussJones007/tarr.pop}
  \item Report bugs at \url{https://github.com/RussJones007/tarr.pop/issues}
}

}
\author{
Russ Jones \href{mailto:RussJones007@gmail.com}{RussJones007@gmail.com}
}
\keyword{package}
